feature : Repository 구성 변경 / 주석 수정 / 비즈니스 로직 변경
