--- src/main/java/com/example/crudproject/common/RedisStringCode.java
@@ -1,5 +1,0 @@
-package com.example.crudproject.common;
-
-public class RedisStringCode {
-    public static final String BOARD_KEY_CODE = "BOARD_";
-}
