+++ src/main/java/com/example/crudproject/common/code/RedisStringCode.java
@@ -1,0 +1,8 @@
+package com.example.crudproject.common.code;
+
+/**
+ * 내부 약속 Redis Key
+ */
+public class RedisStringCode {
+    public static final String BOARD_KEY_CODE = "BOARD";
+}
