+++ src/main/java/com/example/crudproject/common/code/ResponseCode.java
@@ -1,0 +1,12 @@
+package com.example.crudproject.common.code;
+
+/**
+ * 내부 약속 ResponseCode
+ */
+public class ResponseCode {
+    public static final String SUCCESS_CODE = "1000";
+    public static final String ERROR_CODE = "-1000";
+    public static final String OVERLAB_ERROR_CODE = "-1001";
+    public static final String LOGIN_ERROR_CODE = "-1002";
+    public static final String DELETE_ERROR_CODE = "-1003";
+}
