+++ src/main/java/com/example/crudproject/common/config/SecurityConfig.java
@@ -1,0 +1,17 @@
+package com.example.crudproject.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+/**
+ * BCryptPasswordEncoder Config
+ */
+@Configuration
+public class SecurityConfig {
+
+    @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
