--- src/main/java/com/example/crudproject/config/SecurityConfig.java
@@ -1,17 +1,0 @@
-package com.example.crudproject.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-/**
- * BCryptPasswordEncoder Config
- */
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
