--- src/test/java/com/example/crudproject/CrudProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.crudproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CrudProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
