+++ src/test/java/com/example/crudproject/common/CrudProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.crudproject.common;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CrudProjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
